Add shapelib dependency
This will find and add shapelib similarly to how we are finding and
adding googletest with FetchContent.

By default, shapelib will attempt to fetch test data so that it can be
used for tests. This did not work out of the box, and we were unable
to determine how to make it work; so we opted to disable the testing
from shapelib altogether.

The shapelib cmake files do not provide a method to disable testing; it
looks like you should be able to use -DBUILD_TEST=OFF to disable them,
but because the variable is not a CACHE or option() variable, this
doesn't actually work. In order to disable the tests, we are applying
a patch step after fetching the library which will use a cmake
sub-invocation to replace set(BUILD_TEST ON) with set(BUILD_TEST OFF)
throughout the project's cmake file.