Add Doxygen build to CMake
The shpxx-build-docs target is added by default when shpxx is being
built as a top-level project (not being consumed by another project).
In this case, the project internal documentation is also build, but can
be disabled with the SHPXX_BUILD_PRIVATE_DOCS option. This is expected
to be used for public-facing documentation deployed e.g., to a GitHub
pages site for the project.