Fix GCOV_COMMAND not defined after first cmake
Coverage instrumentation currently causes cmake to fail after the
first time that the build files are generated. This is because some
of the variables that FindGCovr sets are CACHE variables, so they
persist, but the GCOV_COMMAND variable (which is used to determine if
gcovr was found) is not, so after the first invocation of cmake, it is
never populate.