shp2csv: Add shapelib example for API comparison
shp2csv will be an example commandline tool to read a shapefile with
shapelib (without shpxx) and dump its geometries and attributes in a CSV
format like

    id,geometry,weight,name
    0,"POINT(0 0)",0.0,"P1"
    1,"POINT(1 1)",0.0,"P2"
    ...

Geometries and string attributes will always quoted. String attributes
will be parsed and any contained quotes will be escaped. The "id" and
"geometry" columns will always be present.

The purpose of the shp2csv example is to get a feel for what using "raw"
shapelib is like, and to provide a baseline to compare the TBD shpxx API
against.