Add concept for IO handling
This new concept defines what functions are required for something to
be used to populate a SAHooks object. The parameters and return types
defined in the concept do not match the SAHooks functions directly, and
this is intentional. The process of wrapping an object to create hooks
should cast types back and forth as appropriate. This is done to make
it easier to write types which can be used to create hooks; they can
use more idiomatic types and not have to worry about the details of
the shapelib typedefs.