Create SAHooks from IoHandler concept
This function can take any type which satisfies the IoHandler concept
and produce a SAHooks object which will defer to the static methods of
that type. Some of the hooks (like atof and error) are not passed
to the underlying handler because they will always be handled the same
way.